impl: set SHELL environment variable
We were assuming that most shells set `SHELL` upon startup, but this is
only true for bash. Usually `SHELL` is only set by `login`, so we need
to set it ourselves when launching the user's shell.

Also remove the `keyEqualsValue` which was only being used by the VSCode
integration. This function escapes environment variable values for a
shell, but we were setting the `code` process's environment with them
directly. This was mangling the values by adding quotes and
backslashes.

Fixes #1454.