devconfig: allow comments and trailing commas
This allows comments and trailing comments in devbox.json as supported
by hujson/JWCC.

```json5
// Top-level comment about the config.
{
	"packages": {
		"go": "1.20",
		
		// Needed to decompress stuff on macOS
		"xz": {
			"version": "latest",
			"platforms": ["aarch64-darwin"],
		},
	},
	"shell": {
		"scripts": {
			/*
			A longer block comment about some build script.
			*/
			"build": "go install ./..",
		},
	},
}
```

As a part of this change, the way the `add`/`rm` commands edit and
marshal devbox.json has changed significantly. We can no longer
overwrite the previous JSON, since that would lose the user's comments
and formatting. There's a new `configAST` type to handle surgical edits
to the JSON so that those things aren't lost. Unmarshalling the config
is still done the same way. The docs in `ast.go` have more details.