[poetry plugin] ensure initHook runs with /bin/sh (#1553)
## Summary

A user is saying that the poetry plugin's init-hook was failing for them
with Fish
shell. While the init-hook's command does work with modern fish shells,
it fails
for older versions of fish. 

In this case, the user is on Ubuntu whose package manager seems to use
an older
version of fish.

Regardless, we can ensure the plugin's init-hooks work for consistently
if we
ensure that they are written as posix scripts executed via /bin/sh. This
PR
seeks to do that.

Fixes #1550


## How was it tested?

```
cd examples/development/poetry/poetry_demo
devbox run run_test
```