internal/lock: canonicalize JSON before hashing
Update the cache key hashing functions so they canonicalize JSON before
hashing. This avoids busting the devbox environment cache whenever
whitespace, comments, or field order changes in devbox.json. It uses the
[experimental JSON v2][1] package to perform the canonicalization.

Because these functions are only used for caching and must be computed
quickly at startup, the hashing algorithm is changed to fnv whereas
before it was a mix of md5 and sha256.

The hashing functions are also now in a separate cachehash package since
they're used for more than just hashing the config.

[1]: https://github.com/golang/go/discussions/63397