[tyson] Add tyson support experiment (#1656)
## Summary

Minimal viable implementation of tyson support. This allows user to
opt-in via feature flag and a `devbox.tson` config. It has no effect for
non-opt

Proof of concept of how this can be used:
https://github.com/jetpack-io/frontend/pull/504 (warning private repo)

To simplify this implementation, it is not possible to use any devbox
command that writes to `devbox.json`. It will fail with an error. We
will also show an error if both configs are present.

Tyson allows more programatic logic in our configs. Specifically:

* Allows importing other configs which adds composability.
* Allows importing of other package manager files (e.g. `package.json`,
`composer.json`, etc which allows you to mimic and enhance their
functionality)
* Allows you to customize hooks, packages, and scripts using host
environment variables. ([Proof of
concept](https://github.com/jetpack-io/axiom/pull/3532))

## How was it tested?

* Tested in the proof of concept link above.
* CICD