[plugins] Implement --disable-plugin flag (#1666)
## Summary

This adds a new `--disable-plugin` flag to `devbox add`. It also adds a
new `disable_plugin` field to the package object in the config. If this
field is true packages with built in plugins will not install the
plugin. If the package does not trigger a plugin, the field is ignored.
The field can be added via flag or manually. It can be removed manually
or by running `devbox add <pkg>` again without the flag.

Other fixes/improvements:

* Moved allowInsecure to `AddOpts` instead of devbox general options.
* Refactored Add() so it no longer violates the linter complexity rule.

Fixes: https://github.com/jetpack-io/devbox/issues/1615

cc: @Lagoja 

## How was it tested?

* Unit tests
* `devbox add python --disable-plugin`