github: add issue template forms (#1764)
Convert our existing issue templates to [forms][1] to make it easier to
get helpful bug/feature info.

Issues created from a template will have a `triage` label applied by
default, indicating that it needs to be looked at by someone from the
team. After an issue has been triaged/validated, this label should be
removed. Hopefully this will make it easier to track issues that are
ready to be worked on.

[1]:
https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests/configuring-issue-templates-for-your-repository#creating-issue-forms