plugins/redis: log to stdout by default
The redis plugin configures redis to log to a file in
`.devbox/virtenv/redis/redis.log`. This causes logs to not show up in
process-compose (with `devbox services up redis`), which relies on
capturing logs from stdout/stderr.

Logging to a file probably predates the process-compose integration back
when there was no other way to view the logs. Today it makes more sense
to default to having logs go to stdout and show up in process-compose.

The trade-off is that logs will go to /dev/null if redis is launched as
a regular daemon (and therefore doesn't have a stdout). In this case,
the user can always uncomment the config line to send logs to a file
again.

Fixes #1727.