[plugins] Implement plugins v2 (#1850)
## Summary

Apologies in advance for the size of this PR. Was not really sure how to
split it up, since it's a fairly large step change.

This PR makes 2 major changes:

* Change how plugin fields are processed. Instead of being processed
independently in several parts of the code (`Env`, `Scripts`, etc) they
are now combined inside of `devconfig` in a simple, predictable way.
* Plugins are now allowed to have all fields that devbox.json have. In
fact the plugin struct now embeds a ConfigFile.

All fields from included plugins are added to the environment. The merge
algorithm is as follows:

* includes are evaluated in order first. Base config is evaluated last.
* Slices are concatenated (init hooks, etc)
* maps are merged with later evaluation replacing previous one on
conflict.
* Packages are deduped by name and concatenated. (i.e. multiple versions
not allowed)

**Known issues/limitations:**

* Github plugins will be slow. When using remote imports, every devbox
command does 1 or more HTTP requests to load the imports.
* central lock file. All dependencies are stored there. 
* While recursive includes are allowed, relative paths are not handled
correctly. That means if project A includes plugin 1 and plugin 1
includes a relative local plugin, it will not work. Including github
plugins recursively works as expected (but will be slow).
* Cycle detection is not implemented. A user can cause infinite include.
* Plugins are json only. They don't support hujson. (This is inline with
current plugins, but we should build support to match devbox.json)

**Additional requires testing:**

- [x] Test create_files works as expected (for both local and github)
- [x] Test process compose works as expected
- [x] Test global push pull (config changes may have affected that)
- [x] Ensure that we are able to `add` a package that is already added
by plugin.


Follow up cleanup:

* Remove tyson support (this will simplify config code)
* Plugin manager can be removed or greatly reduced.
* Fix some of the limitations (relative paths, cycle detection)
* Cache included remote plugins for performance (and to avoid github
limits).

## How was it tested?

* New example tests
* Manual testing
* Existing tests