[plugins] support reading (and discarding) comments plugin.json files
Fixes #1911

This PR cleans the json-file-content of comments before processing it as "json content" i.e. we discard the comments prior to invoking existing logic for ingesting the `plugin.json` file.

This PR relies on Devbox not programmatically modifying the `plugin.json`. For now, users hand edit the `plugin.json` files. This is different from `devbox.json` which we do programmatically modify due to `devbox add`, `devbox rm` and similar commands, and so need to keep track of the AST.

To test, I added comments to the following:
1. local plugin example
2. a builtin plugin (python)

I did NOT test remote Github plugin. But I tried to modify its code path. Did i get it right?