[offline] Allow devbox to work offline if DNS fails (#1985)
## Summary

Fixes https://github.com/jetify-com/devbox/issues/1977

(with another hack).

Unfortunately the way we structure our flake makes it really difficult
to remove our network requests (when state is stale). Ideally, the flake
can be built off of local only nix store references, but that's a much
bigger project.

In the meantime, this is a quick hack that fixes issue above.

## How was it tested?

Added synthetic error that returns `connection refused` in the message.
Ensures I was still able to run/shell with a stale state.