nixcache: automatically configure nix daemon
Before installing packages, check to see if the Nix daemon is already
configured to use a Devbox cache when one is available. The user no
longer needs to explicitly call `devbox cache configure`, however the
command is still there (and hidden) to make sudo escalation possible.

There are a few known limitations in this initial implementation:

- We don't keep track of "no" responses to sudo, so Devbox will bug the
  user every time if they refuse.
- The Nix daemon needs to restart after adding a trusted user. We
  currently only support auto-restarting the daemon on macOS and assume
  its managed with launchd.
- We sleep rather than poll for the daemon to restart. This only happens
  when a trusted user is added, but it does make things a little
  flakier.