[docker] Added fix to run image release on tags creation + backfill action (#2118)
## Summary
Github action for releasing docker images to dockerhub didn't run when
we created a new tag (0.11.0) this fixes that issue.
Plus I added a new github action that can only be triggered manually. It
takes inputs for tag/version name and releases docker image to dockerhub
at that specific tag. This is to backfill older versions of devbox to
dockerhub as well.

## How was it tested?
Will have to be tested via github actions run on a tag creation. But
backfill will be tested manually