internal/nix: log all command starts and exits
Write logs whenever a Nix command is about to start and after it exits.
Make `cmd` a `slog.LogValuer` so that it writes structured log
attributes grouped under a `cmd` prefix.

Example logs and attributes (line breaks added for readability):

	msg="nix command starting"
		cmd.args="nix --extra-experimental-features ca-derivations --option experimental-features 'nix-command flakes fetch-closure' eval --impure --raw --expr builtins.currentSystem"
		cmd.path=/nix/var/nix/profiles/default/bin/nix

	msg="nix command exited"
		cmd.args="nix --extra-experimental-features ca-derivations --option experimental-features 'nix-command flakes fetch-closure' eval --impure --raw --expr builtins.currentSystem"
		cmd.path=/nix/var/nix/profiles/default/bin/nix
		cmd.pid=68672
		cmd.code=0
		cmd.dur=34.296041ms

Command exits are always logged at the debug level, even if the command
exits with a non-zero exit code. This is because some commands might be
expected to fail (for example, when checking if a path exists in the
store).