testscripts: python tests for numpy and psycopg2 (#2214)
Add tests that install Python with Devbox and then run scripts that import numpy
and psycopg2. They currently fail on Linux due to linker errors.

To run the tests on Linux with Docker, run:

	devbox run -e DEVBOX_RUN_FAILING_TESTS=1 docker-testscripts -test.run '^TestScripts$/python'

The new docker-testscripts script in devbox.json runs testscripts in a Docker
container to make it easier to test various Linux distros and architectures.