testscripts: better caching with Docker
Mount a Docker volume to /nix when running `docker-testscripts` to speed
up tests. Also set `XDG_CACHE_HOME=/nix/cache` so that the tarball and
eval cache persist to the volume. Update `setupTestEnv` to share the pip
cache between tests (the same way we do with Nix's cache).

This cuts the duration of python linker tests in half (down from about 1
minute to about 30s).