patchpkg: `lookPath` helper to find nix binaries
When running inside a Nix derivation the PATH is completely empty, so
functions like `exec.Command("bash")` don't work. To make dealing with
this easier, add a helper function that looks for an environment
variable named after the command and returns `$cmd/bin/cmd` (e.g.,
`bash` becomes `$bash/bin/bash`). This will automatically find
build-time dependencies that are added as attributes to the derivation.