[services] Let users set the port for process-compose (#2299)
## Summary

Users can set the port that process-compose runs on when they use
`devbox services`. If they do not specify a port, we default to randomly
picking an open port with low risk of conflict

Users can specify their process-compose port by:

1. Passing the `--pcport, -p` flag when running `devbox services up`. 
2. Setting the `PC_PORT_NUM` environment variable. This variable can be
set in the projects `devbox.json` if you want to always use the same
port for process-compose. This is overridden by the `--pcport` flag.

## How was it tested?

In the MySQL example: 

1. When I run `devbox services up -b -p 8080`, or
2. When I run `PC_PORT_NUM devbox services up -b`,

I can verify that process-compose is running on my 8080 with `curl
localhost:8080/processes` and seeing the list of expected processes.