[telemetry] Don't log user errors (#2305)
## Summary

We're running out of Sentry events and it looks like the culprit is
devbox overlogging, including a lot of user errors that should not be
logged.

This PR https://github.com/jetify-com/devbox/pull/844 removed code that
avoided logging user errors. I think this was a bug? If programmer wants
a user error to be logged, they should use `WithLoggedUserMessage`
instead.

## How was it tested?

Untested